Module rustpython_compiler::symboltable

Structs

Symbol
A single symbol in a table. Has various properties such as the scope of the symbol, and also the various uses of the symbol.

SymbolTable
Captures all symbols in the current scope, and has a list of subscopes in this scope.

SymbolTableError

Enums

SymbolScope
Indicator for a single symbol what the scope of this symbol is. The scope can be unknown, which is unfortunate, but not impossible.

SymbolTableType

Functions

make_symbol_table

statements_to_symbol_table